naprawione dodanie floatow jako ilosc

+++ backend/src/main/java/com/zjadbyco/models/Food.java
@@ -1,7 +1,7 @@
 package com.zjadbyco.models;
 
-public record Food(String id, String name, int quantity, String unit, String date) {
-    public Food(String id, String name, int quantity, String unit) {
+public record Food(String id, String name, float quantity, String unit, String date) {
+    public Food(String id, String name, float quantity, String unit) {
         this(id, name, quantity, unit, null);
     }
 }
